feat: use states for lists management
