+++ src/models/Campaign.ts
@@ -16,8 +16,6 @@ export interface ICampaignUpdatePayload {
 }
 
 export interface ICampaign {
-  startAt: any;
-  endAt: any;
   id?: number;
   promotionId: number;
   subject: string;
