+++ src/pages/campaign/index.tsx
@@ -1,14 +1,93 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import Sidebar from '@/components/layout/Sidebar';
 import { CampaignCreateForm } from '@/components/form/Campaign.form';
 import CampaignList from '@/components/list/CampaignList';
 import Navbar from '@/components/layout/Navbar';
+import { ICampaign } from '@/models/Campaign';
+import { IUser } from '@/models/User';
+import { IPromotion } from '@/models/Promotions';
 
 export default function Campaign() {
   const [isShown, setIsShown] = useState(false);
+  const [campaigns, setCampaigns] = useState<ICampaign[]>([]);
+  const [promotions, setPromotions] = useState<IPromotion[]>([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState(false);
+
+  const loadUser = async (): Promise<IUser> => {
+    const userUuid = localStorage.getItem('userUuid');
+    const options = {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    };
+
+    try {
+      const response = await fetch(`/api/user/${userUuid}`, options);
+      const data = await response.json();
+      return data;
+    } catch (error) {
+      setError(true);
+    }
+  };
+
+  const loadCampaigns = async (): Promise<void> => {
+    setIsLoading(true);
+
+    const user = await loadUser();
+    const options = {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    };
+
+    try {
+      const response = await fetch(
+        `/api/shop/${user.shop.id}/campaigns`,
+        options,
+      );
+      const data = await response.json();
+      const sortedList: ICampaign[] = data.sort(
+        (a, b) => a.updatedAt - b.updatedAt,
+      );
+      setCampaigns(sortedList);
+    } catch (error) {
+      setError(true);
+    }
+
+    setIsLoading(false);
+  };
+
+  const loadPromotions = async (): Promise<void> => {
+    const options = {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    };
+    const user = await loadUser();
+    try {
+      const response = await fetch(
+        `/api/shop/${user.shop.id}/promotion`,
+        options,
+      );
+      const data = await response.json();
+      setPromotions(data);
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
   const handleClick = (): void => {
     setIsShown((current) => !current);
   };
+
+  useEffect(() => {
+    loadPromotions();
+    loadCampaigns();
+  }, []);
 
   return (
     <div className="p-4 sm:p-6 lg:p-8 min-h-screen">
@@ -35,10 +114,20 @@ export default function Campaign() {
       </div>
       {isShown && (
         <div>
-          <CampaignCreateForm />
+          <CampaignCreateForm
+            setShown={setIsShown}
+            setCampaigns={setCampaigns}
+            campaigns={campaigns}
+            promotions={promotions}
+          />
         </div>
       )}
-      <CampaignList />
+      <CampaignList
+        campaigns={campaigns}
+        setCampaigns={setCampaigns}
+        isLoading={isLoading}
+        error={error}
+      />
     </div>
   );
 }
